[BE] refactor : enums 패키지로 Enum 값 관리
