+++ backend/src/main/java/booki_today/domain/regular_moim/enums/RegularMoimCategory.java
@@ -1,4 +1,4 @@
-package booki_today.domain.regular_moim;
+package booki_today.domain.regular_moim.enums;
 
 // todo 정기 모임 카테고리 세부사항 추가
 public enum RegularMoimCategory {
