+++ backend/src/main/java/booki_today/domain/single_moim/enums/SingleMoimCategory.java
@@ -1,4 +1,4 @@
-package booki_today.domain.single_moim;
+package booki_today.domain.single_moim.enums;
 
 // todo 단기 모임 카테고리 세부사항 추가
 public enum SingleMoimCategory {
