+++ backend/src/main/java/booki_today/persistence/entity/member/MemberJpaEntity.java
@@ -1,6 +1,6 @@
 package booki_today.persistence.entity.member;
 
-import booki_today.domain.member.Gender;
+import booki_today.domain.member.enums.Gender;
 import booki_today.global.annotation.Association;
 import booki_today.global.base_entity.BaseTimeEntity;
 import jakarta.persistence.*;
