+++ backend/src/main/java/booki_today/persistence/entity/regular_moim/moim_schedule/MoimScheduleJpaEntity.java
@@ -1,6 +1,6 @@
 package booki_today.persistence.entity.regular_moim.moim_schedule;
 
-import booki_today.domain.regular_moim.DayOfWeek;
+import booki_today.domain.regular_moim.enums.DayOfWeek;
 import booki_today.global.annotation.Association;
 import booki_today.global.base_entity.BaseTimeEntity;
 import jakarta.persistence.*;
