+++ backend/src/main/java/booki_today/persistence/entity/regular_moim/regular_moim/RegularMoimJpaEntity.java
@@ -1,6 +1,6 @@
 package booki_today.persistence.entity.regular_moim.regular_moim;
 
-import booki_today.domain.regular_moim.RegularMoimCategory;
+import booki_today.domain.regular_moim.enums.RegularMoimCategory;
 import booki_today.global.annotation.Association;
 import booki_today.global.base_entity.BaseTimeEntity;
 import jakarta.persistence.*;
